Refactor action to callback
