+++ bokeh/models/__init__.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import
 
 # This file is excluded from flake8 checking in setup.cfg
 
-from .actions import *
+from .callbacks import *
 from .axes import *
 from .formatters import *
 from .glyphs import *
