+++ bokeh/models/sources.py
@@ -6,7 +6,7 @@ from ..properties import Any, Int, String, Instance, List, Dict, Either, Bool, E
 from ..validation.errors import COLUMN_LENGTHS
 from .. import validation
 from ..util.serialization import transform_column_source_data
-from .actions import Callback
+from .callbacks import Callback
 from bokeh.deprecate import deprecated
 
 class DataSource(PlotObject):
