+++ bokeh/models/widgets/inputs.py
@@ -6,7 +6,7 @@ from __future__ import absolute_import
 import six
 
 from ...properties import Bool, Int, Float, String, Date, RelativeDelta, Enum, List, Dict, Tuple, Either, Instance
-from ..actions import Callback
+from ..callbacks import Callback
 from ..widget import Widget
 
 class InputWidget(Widget):
