+++ bokehjs/src/coffee/callback/open_url.coffee
