+++ sphinx/source/docs/user_guide/source_examples/interaction_callbacks_for_selections.py
@@ -1,5 +1,5 @@
 from random import random
-from bokeh.models import Callback, ColumnDataSource
+from bokeh.models import CustomJS, ColumnDataSource
 from bokeh.plotting import hplot, figure, output_file, show
 
 output_file("callback.html")
@@ -16,7 +16,7 @@ p2 = figure(plot_width=400, plot_height=400, x_range=(0, 1), y_range=(0, 1),
             tools="", title="Watch Here")
 p2.circle('x', 'y', source=s2, alpha=0.6)
 
-s1.callback = Callback(args=dict(s2=s2), code="""
+s1.callback = CustomJS(args=dict(s2=s2), code="""
         var inds = cb_obj.get('selected')['1d'].indices;
         var d1 = cb_obj.get('data');
         var d2 = s2.get('data');
